cloud_functions/download_pipeline/requirements.txt
functions-framework==3.*
pandas
yt_dlp
google-cloud_bigquery
pandas-gbq
google-cloud-storage
opencv-python